requirements.txt
fastapi~=0.115.0
Pillow~=10.4.0
ultralytics~=8.2.101
timm~=1.0.9
PyJWT~=2.9.0
passlib~=1.7.4
torchvision~=0.19.1
dagshub~=0.3.36
mlflow~=2.16.2
pandas~=2.2.3
evidently~=0.4.37
huggingface-hub~=0.25.1